% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fomir-package.R
\docType{package}
\name{Fomir-package}
\alias{Fomir}
\alias{Fomir-package}
\title{Fomir: Forest Operations Models in R}
\description{
This package provide implementation of forest operations models for estimating time consumtion, costs, etc in forest operations.
}
\author{
\strong{Maintainer}: Helmer Belbo \email{helmer.belbo@nibio.no}

}
\keyword{internal}
