% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_residues_forwarder.R
\name{lr_forw}
\alias{lr_forw}
\title{Logging residues forwarder model
Time consumption model: Nurmi 2007, Recovery of LR for energy from spruce stands.
Fuel consumption model: Brunberg 2012, Bränsleförbrukningen hos skogsmaskiner 2012.}
\usage{
lr_forw(species, ForwD_km)
}
\arguments{
\item{species}{numeric code: 1 = gran, 2 = furu, 3 = lauv, 4 = andre}

\item{ForwD_km}{numeric: forwarding distance (one way)}
}
\value{
single number, the forwarder cost (NOK) per m3solid logging residues
}
\description{
Logging residues forwarder model
Time consumption model: Nurmi 2007, Recovery of LR for energy from spruce stands.
Fuel consumption model: Brunberg 2012, Bränsleförbrukningen hos skogsmaskiner 2012.
}
\examples{
lr_forw(species = 1, ForwD_km = 0.4)
}
