% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heureka brunberg.R
\name{t_forw}
\alias{t_forw}
\title{t_forw}
\usage{
t_forw(stands_df, forw_size = "large", modelversion = "Talbot16")
}
\arguments{
\item{stands_df}{A data frame of stands for which forw time consumption is to
be estimated.
Mandatory variables in stands_df, with default values:
  v = 0.3; mean volume m3fub of the harvested trees in the stand
  Nharv = 1000; Number of trees per hectare
 Optional variables, of which default values is set as first value:
  nbAssortments = 4; Number of qualities and species. Afect t7 ForwarderSortingTime
  treatment = one of : "clearcutting", "thinning"
  L = 2, slope class (1:5) where 5 is steep
  Y = 2, surface structure (1:5) where 5 is uneven
  D = 300, Terrain transport distance (one way average)}

\item{forw_size}{can be "small", "medium", "large"
Should also take "xl" and "xxl" but not implemented}

\item{modelversion}{one of "Brunberg" | "Talbot16"}
}
\value{
a data frame with forwarder time consumption (minutes per hectare)
}
\description{
t_forw
}
\examples{
t_forw(stands_df = heureka_testdat_forwarding, forw_size = "medium", modelversion = "Brunberg") \%>\% glimpse()
t_forw(stands_df = heureka_testdat_forwarding, forw_size = "medium", modelversion = "Talbot16") \%>\% glimpse()
t_forw(stands_df = heureka_testdat_thinning) \%>\% glimpse()
t_forw(stands_df = heureka_testdata_clearcutting) \%>\% glimpse()
t_forw(stands_df = testdf_volume) \%>\% glimpse()

}
