% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forw_bb.R
\name{t_forw_bb}
\alias{t_forw_bb}
\title{t_forw_bb}
\usage{
t_forw_bb(
  stands_df,
  forw_size = "large",
  modelversion = c("Brunberg04", "Talbot16")[2],
  treatment = c("thinning", "clearcutting")[2],
  harvest_strength_pct = 100,
  verbose = FALSE
)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics according to the
variables listed under heading "3 Variables" in the Heureka time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf

Mandatory variables in stands_df, with default values:
  v = 0.3; mean volume m3fub of the harvested trees in the stand
  Nharv = 1000; Number of trees per hectare
 Optional variables, of which default values is set as first value:
  nbAssortments = 4; Number of qualities and species. Afect t7 ForwarderSortingTime
  treatment = one of : "clearcutting", "thinning"
  L = 2, slope class (1:5) where  1 is flat (0-10%), 2 is 10-20%, 3 is 20-33%, 4 is 33-50%, 5 is >50%
  Y = 2, surface structure (1:5) where 5 is uneven
  D = 300, Terrain transport distance (one way average)}

\item{forw_size}{can be "small", "medium", "large"
Should also take "xl" and "xxl" but not implemented}

\item{modelversion}{one of "Brunberg" | "Talbot16"}

\item{treatment}{one of "thinning" | "clearcutting"}

\item{harvest_strength_pct}{share of basal area to be cut. Range 0-100}

\item{verbose}{if TRUE: more intermediate time consumption sstimates will be returned}
}
\value{
a data frame with forwarder time consumption (minutes per hectare)
\strong{variables in output df:}
\describe{
\item{forw_G15h.ha}{Forwarder G15 hours per hectare}}
}
\description{
#' @description
`t_forw_bb()` estimate forwarder time consumption in forest operations, according
to models presented in the swedish Heureka description,
and a later modified version by Bruce Talbod (NIBIO Norway)
}
\examples{
t_forw_bb(stands_df = testdata_heureka_forwarding, verbose = TRUE) \%>\% dplyr::glimpse()
t_forw_bb(stands_df = testdata_heureka_thinning, verbose = TRUE) \%>\% dplyr::glimpse()
t_forw_bb(stands_df = testdata_heureka_clearcutting, verbose = TRUE) \%>\% dplyr::glimpse()
}
