% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_forw_bledn_joh15.R
\name{t_forw_bledn_joh15}
\alias{t_forw_bledn_joh15}
\title{t_forw_bledn_joh15}
\usage{
t_forw_bledn_joh15(stands_df, treatmentnumber = 2, verbose = FALSE)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics to be used in the
function. Must include:
'Stems_ha' Number of stems per hectare,
'v' mean volume m3fub of the harvested trees in the stand.
Can also include:
Nharv number of harvested stems per hectare
'p_init_spruce' proportion of spruce before harvest
'L' slope class (1:5) where  1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\%
'Y' surface structure (1:5) where 5 is uneven}

\item{treatmentnumber}{one of: 1, 2. 1 is first (initiating bledning from even-aged stand), 2 is later bledning where striproads exist}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A [tibble()].
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
 harv_G15min.tree (harvest time, G15 minutes per tree)
 harv_G15h.ha (harvest time, G15 hours per hectare)
harv_G15min.m3 (harvest time, G15 minutes per m3)
harv_m3.G15h (productivity, m3 per G15-hour)
}
\description{
`t_forv_bledn_joh15()` estimate harvester time consumption in forest operations, according
to models presented in the swedish Heureka description,
and a later modified version by Bruce Talbod (NIBIO Norway), and a
model for "bledningshogst" by Jonsson in 2015
}
\details{
variables to provide is listed under heading "3 Variables" in the Heureka
time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf
Clearcut harvester model is the Heureka clearcut model, being a
slightly modified version of Brunberg 1995.
Thinning harvester model is the Heureka clearcut model, being a
slightly modified version of Brunberg1997
}
\examples{
t_forw_bledn_joh15(stands_df = testdata_heureka_clearcutting, verbose = TRUE) \%>\% dplyr::glimpse()
t_forw_bledn_joh15(stands_df = testdata_heureka_thinning, verbose = TRUE) \%>\% dplyr::glimpse()
t_forw_bledn_joh15(stands_df = testdata_sensblt_v_Y_L_D, verbose = TRUE) \%>\% dplyr::glimpse()
t_forw_bledn_joh15(stands_df = testdata_fjeld, verbose = TRUE) \%>\% dplyr::glimpse()
}
