% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heureka brunberg.R
\name{t_harv}
\alias{t_harv}
\title{t_harv}
\usage{
t_harv(stands_df, modelversion = "Talbot16")
}
\arguments{
\item{stands_df}{A data frame having stand characteristics according to the
variables listed under heading "3 Variables" in the Heureka time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf

Mandatory variables in stands_df:
  v = mean volume m3fub of the harvested trees in the stand.
  Nharv = Number of trees per hectare
Optional variables, of which default values is set as first value here in the description
 for clearcutting & thinning:
 L = 2, slope class (1:5) where 5 is steep
 Y = 2, surface structure (1:5) where 5 is uneven
 treatment = one of : "clearcutting", "thinning"
 Nres = 0; Number of trees left per hectare (used in thinning time calc)
 thinningsystem = one of: striproad, striproad_with_midfield_machine, striproad_with_midfield_chainsaw
 thinningnumber = one of: 1, 2,
 rd = relative diameter in thinning = 0.9, i.e. mean dia harvested / mean dia residual trees. BA weighted mean diameters.
 p_init_spruce = 0.9, proportion of spruce before harvest
 Pharv_broadLeaves = .9 proportion broadleaved trees of harvested trees
 W = 16, width of strips between striproads (excluding striproad)}

\item{modelversion}{one of "Talbot16"|"Brunberg"}
}
\value{
a data frame having the estimated time consumption cmin per tre (ttot), and also:
harv_G15min.tree (harvest time, G15 minutes per tree)
harv_G15h.ha (harvest time, G15 hours per hectare)
harv_G15min.m3 (harvest time, G15 minutes per m3)
harv_m3.G15h (productivity, m3 per G15-hour)
}
\description{
t_harv
}
\examples{
t_harv(stands_df = heureka_testdata_clearcutting, modelversion = "Brunberg") \%>\% dplyr::glimpse()
t_harv(stands_df = heureka_testdata_clearcutting, modelversion = "Talbot16") \%>\% dplyr::glimpse()
t_harv(stands_df = heureka_testdata_thinning, modelversion = "Brunberg") \%>\% dplyr::glimpse()
t_harv(stands_df = heureka_testdata_thinning, modelversion = "Talbot16") \%>\% dplyr::glimpse()
t_harv(stands_df = stand_data_norsk_prodnorm) \%>\% dplyr::glimpse()
t_harv(stands_df = heureka_testdata_clearcutting) \%>\% dplyr::glimpse()

}
