% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_harvester.R
\name{t_harv_bb}
\alias{t_harv_bb}
\title{t_harv_bb}
\usage{
t_harv_bb(
  stands_df,
  Nharv = 700,
  v = 0.4,
  modelversion = "Brunberg",
  treatment = "clearcutting",
  thinningsystem = NA_character_,
  thinningnumber = NA_real_,
  rd = NA_real_,
  Nres = NA_real_,
  p_init_spruce = 1,
  Pharv_broadLeaves = 0,
  L = 2,
  Y = 2,
  W = 16,
  verbose = FALSE
)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics to be used in the
function. Can include all or any or none of the variables below. Variables in
stands_df will override variables in the function call.}

\item{Nharv}{Number of trees per hectare'}

\item{v}{mean volume m3fub of the harvested trees in the stand.}

\item{modelversion}{one of "Talbot16" or "Brunberg"}

\item{treatment}{one of : "clearcutting", "thinning"}

\item{thinningsystem}{one of: striproad, striproad_with_midfield_machine, striproad_with_midfield_chainsaw}

\item{thinningnumber}{one of: 1, 2,}

\item{rd}{relative diameter in thinning,
i.e. mean dia harvested / mean dia residual trees. BA weighted mean diameters.}

\item{Nres}{Number of trees left per hectare (used in thinning time calc)}

\item{p_init_spruce}{proportion of spruce before harvest}

\item{Pharv_broadLeaves}{proportion broadleaved trees of harvested trees}

\item{L}{slope class (1:5) where  1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\%}

\item{Y}{surface structure (1:5) where 5 is uneven}

\item{W}{width of strips between striproads (excluding striproad)}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A [tibble()].
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
 harv_G15min.tree (harvest time, G15 minutes per tree)
 harv_G15h.ha (harvest time, G15 hours per hectare)
harv_G15min.m3 (harvest time, G15 minutes per m3)
harv_m3.G15h (productivity, m3 per G15-hour)
}
\description{
`t_harv_bb()` estimate harvester time consumption in forest operations, according
to models developed by Torbjörn Brunberg (Skogforsk Sweden) and
later modified version by Bruce Talbod (NIBIO Norway)
}
\details{
variables to provide is listed under heading "3 Variables" in the Heureka
time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf
}
\examples{
t_harv_bb(stands_df = testdata_heureka_clearcutting, modelversion = "Brunberg") \%>\% dplyr::glimpse()
t_harv_bb(stands_df = testdata_heureka_clearcutting, modelversion = "Talbot16") \%>\% dplyr::glimpse()
t_harv_bb(stands_df = testdata_heureka_thinning) \%>\% dplyr::glimpse()
t_harv_bb(stands_df = testdata_clearcut_Talbot16, verbose = TRUE) \%>\% dplyr::glimpse()

}
