% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_harv_bledn_joh15.R
\name{t_harv_bledn_joh15}
\alias{t_harv_bledn_joh15}
\title{t_harv_bledning_joh15}
\usage{
t_harv_bledn_joh15(stands_df, treatmentnumber = 2, verbose = FALSE)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics to be used in the
function. Must include:
\itemize{
\item 'Stems_ha' Number of stems per hectare,
\item 'v' mean volume m3fub of the (harvestable) trees in the stand.
}}

\item{treatmentnumber}{one of: 1, 2. 1 is first (initiating bledning from even-aged stand), 2 is later bledning where striproads exist}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A [tibble()].
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
 harv_G15min.tree (harvest time, G15 minutes per tree)
 harv_G15h.ha (harvest time, G15 hours per hectare)
harv_G15min.m3 (harvest time, G15 minutes per m3)
harv_m3.G15h (productivity, m3 per G15-hour)
}
\description{
`t_harv_bledn_joh15()` estimate harvester time consumption in forest operations, according
to models presented "bledningshogst" by Jonsson in 2015
}
\examples{
t_harv_bledn_joh15(stands_df = testdata_heureka_clearcutting, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_bledn_joh15(stands_df = testdata_heureka_thinning, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_bledn_joh15(stands_df = testdata_sensblt_v_Y_L_D, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_bledn_joh15(stands_df = testdata_clearcut_forw, verbose = TRUE) \%>\% dplyr::glimpse()
}
