% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_harv_ccf_fj94.R
\name{t_harv_ccf_fj94}
\alias{t_harv_ccf_fj94}
\title{t_harv_ccf_fj94}
\usage{
t_harv_ccf_fj94(
  stands_df,
  Stems_ha = 1200,
  v = 0.4,
  treatment = c("patch_harvest", "single_tree_selection")[1],
  harvest_strength_pct = 45,
  patch_size_ha = 0.25,
  sr_w = 4,
  striproad_exist = FALSE,
  Y = 1,
  L = 1,
  verbose = FALSE
)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics to be used in the
May include all the parameters set in the function call except "verbose"}

\item{Stems_ha}{Number of stems per hectare,}

\item{v}{mean volume m3fub of the trees in the stand.}

\item{treatment}{one of :  "single_tree", "patch"}

\item{harvest_strength_pct}{share of basal area to be cut. Range 0-100}

\item{patch_size_ha}{size (hectare) of each harvested patch in patch-based harvesting}

\item{sr_w}{strip road width in meter}

\item{striproad_exist}{TRUE/FALSE if striproad allready exist}

\item{Y}{surface structure (1:5) where 5 is uneven}

\item{L}{slope class (1:5) where  1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\%}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A tibble.
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
\itemize{
\item harv_G15min.tree (harvest time, G15 minutes per tree)
\item harv_G15h.ha (harvest time, G15 hours per hectare)
\item  harv_G15min.m3 (harvest time, G15 minutes per m3)
\item  harv_m3.G15h (productivity, m3 per G15-hour)
}
}
\description{
`t_harv_ccf_fj94()` estimates harvester time consumption for felling and processing trees using
Cut-to-length single grip harvester in selective and patch cutting, i.e. two continious cover approaches
}
\details{
The study was done using a Valmet 901, 83 kW and 11 tonne, 9.6m boom reach.
THe study was done on even surface, good to medium ground conditions,
flat terrain, with some boulders.
The stands were spruce dominated, site index G11 - G15,
stand density was 1800 - 2100 stems/ha, 163 - 308 m3/ha.
}
\examples{
t_harv_ccf_fj94(stands_df = testdata_heureka_thinning, treatment = "patch_harvest", harvest_strength_pct = 25, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_heureka_clearcutting, treatment = "patch_harvest", harvest_strength_pct = 45,  verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_heureka_clearcutting, treatment = "patch_harvest", harvest_strength_pct = 45, striproad_exist = TRUE, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_heureka_clearcutting, treatment = "single_tree_selection", harvest_strength_pct = 45,  verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_heureka_clearcutting, treatment = "single_tree_selection", harvest_strength_pct = 45, striproad_exist = TRUE,  verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_fjeld,  treatment = "single_tree_selection", harvest_strength_pct = 45, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_ccf_fj94(stands_df = testdata_fjeld,  treatment = "patch_harvest", harvest_strength_pct = 45, verbose = TRUE) \%>\% dplyr::glimpse()
}
