% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_harv_clearfell_bb.R
\name{t_harv_clearfell_bb}
\alias{t_harv_clearfell_bb}
\title{t_harv_clearfell_bb}
\usage{
t_harv_clearfell_bb(
  stands_df = NULL,
  Stems_ha = 1000,
  v = 0.5,
  L = 2,
  Y = 2,
  modelversion = c("Brunberg95", "Talbot16", "Brunberg07")[1],
  verbose = FALSE
)
}
\arguments{
\item{stands_df}{A data frame having stand characteristics to be used in the
function.
May include all the parameters set in the function call except "verbose"}

\item{Stems_ha}{Number of stems per hectare,}

\item{v}{mean volume m3fub of the trees in the stand.}

\item{L}{slope class (1:5) where  1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\%}

\item{Y}{surface structure (1:5) where 5 is uneven}

\item{modelversion}{one of "Brunberg95" or "Talbot16"}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A tibble.
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
\itemize{
\item harv_G15min.tree (harvest time, G15 minutes per tree)
\item harv_G15h.ha (harvest time, G15 hours per hectare)
\item  harv_G15min.m3 (harvest time, G15 minutes per m3)
\item  harv_m3.G15h (productivity, m3 per G15-hour)
}
}
\description{
`t_harv_clearfell_bb()` estimate harvester time consumption in forest operations, according
to models presented in the swedish Heureka description,
and a later modified version by Bruce Talbod (NIBIO Norway), and a
model for "bledningshogst" by Jonsson in 2015
}
\details{
variables to provide is listed under heading "3 Variables" in the Heureka
time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf
Clearcut harvester model is the Heureka clearcut model, being a
slightly modified version of Brunberg 1995.
Thinning harvester model is the Heureka clearcut model, being a
slightly modified version of Brunberg1997
}
\examples{
t_harv_clearfell_bb(stands_df = testdata_heureka_clearcutting,  verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_clearfell_bb(stands_df = testdata_heureka_clearcutting, modelversion = "Talbot16", verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_clearfell_bb(stands_df = testdata_heureka_clearcutting, modelversion = "Brunberg07", verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_clearfell_bb(stands_df = testdata_heureka_thinning, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_clearfell_bb(stands_df = testdata_clearcut_forw, modelversion = "Talbot16",  verbose = TRUE) \%>\% dplyr::glimpse()
}
