% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_harv_thinning_bb.R
\name{t_harv_thinning_bb}
\alias{t_harv_thinning_bb}
\title{t_harv_thinning_bb}
\usage{
t_harv_thinning_bb(
  stands_df = NULL,
  Stems_ha = 1200,
  v = 0.4,
  harvest_strength_pct = 30,
  p_init_spruce = 0.95,
  L = 2,
  Y = 2,
  modelversion = "Brunberg97",
  Pharv_broadLeaves = 0,
  thinningsystem = c("striproad_with_midfield_machine",
    "striproad_with_midfield_chainsaw", "striproad")[3],
  thinningnumber = 1,
  rd = 0.9,
  sr_sp = 20,
  sr_w = 4,
  verbose = FALSE
)
}
\arguments{
\item{stands_df}{A data frame having stand and operational characteristics to be used in the
function. May include all the parameters set in the function call except "verbose"}

\item{Stems_ha}{Number of stems per hectare,}

\item{v}{mean volume m3fub of the trees in the stand.}

\item{harvest_strength_pct}{share of basal area to be cut. Range 0-100}

\item{p_init_spruce}{proportion of spruce before harvest}

\item{L}{slope class (1:5) where  1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\%}

\item{Y}{surface structure (1:5) where 5 is uneven}

\item{modelversion}{one of "Brunberg95" or"Brunberg07", "Talbot16"}

\item{Pharv_broadLeaves}{proportion broadleaved trees of harvested trees}

\item{thinningsystem}{one of: striproad, striproad_with_midfield_machine, striproad_with_midfield_chainsaw}

\item{thinningnumber}{one of: 1, 2, (Applies also to bledningshogst, 1 is first, 2 is later bledning where striproads exist)}

\item{rd}{relative diameter in thinning,
i.e. mean dia harvested / mean dia residual trees. BA weighted mean diameters.}

\item{sr_sp}{striproad spacing centre to centre}

\item{sr_w}{striproad widt: 2 * distance striporoad centreline to nearest tree}

\item{verbose}{if TRUE: returning all time components from time consumption models}
}
\value{
A tibble.
If verbose TRUE; estimates from sub calculations are also present in the table.
If verbose FALSE; time consumption per production unit and productivity in
  the following variables:
\itemize{
\item harv_G15min.tree (harvest time, G15 minutes per tree)
\item harv_G15h.ha (harvest time, G15 hours per hectare)
\item  harv_G15min.m3 (harvest time, G15 minutes per m3)
\item  harv_m3.G15h (productivity, m3 per G15-hour)
}
}
\description{
`t_harv_thinning_bb()` estimates harvester time consumption in forest operations, according
to models presented in the swedish Heureka description
}
\details{
Variables to provide is listed under heading "3 Variables" in the Heureka
time consumption functions description
https://www.heurekaslu.se/w/images/8/8c/Time_consumption_harvester_and_forwarder.pdf
The thinning harvester model is the Heureka clearcut model, being a
slightly modified version of Brunberg1997
NB: variables listed as arguments may also be provided in the "stands_df" dataset.
}
\examples{
t_harv_thinning_bb(stands_df = testdata_heureka_thinning, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_thinning_bb(stands_df = testdata_clearcut_forw, verbose = TRUE) \%>\% dplyr::glimpse()
t_harv_thinning_bb(stands_df = testdata_heureka_clearcutting,  verbose = TRUE) \%>\% dplyr::glimpse()
}
