% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata_fjeld}
\alias{testdata_fjeld}
\title{Test data set for testing Fjeld 94 bledningshogst}
\format{
data.frame() with the following content:
\describe{
 \item{v}{numeric; mean stem volume m3fub}
 \item{Stems_ha}{Number of stems per hectare}
}
}
\usage{
testdata_fjeld
}
\description{
Test data set for testing Fjeld 94 bledningshogst
}
\keyword{datasets}
