% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata_heureka_clearcutting}
\alias{testdata_heureka_clearcutting}
\title{Test data set for Brunberg's clearcutting model}
\format{
data.frame() with the following content:
\describe{
 \item{v}{numeric; mean stem volume m3fub}
 \item{Nharv}{numeric; Number of harvested stems per hectare}
 \item{Y}{numeric; Surface structure (sv: Ytstruktur)}
 \item{L}{numeric; Slope class (sv: Lutning), Definition according “SkogForsk
 terrängtypschema” where 1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\% }
 \item{treatment}{text; "clearcutting" (or "thinning")  }
 \item{solution_hr}{numeric; expected model output, to check if Brunberg's
 clearcutting  model is doing correct estimates on the dataset}
}
}
\usage{
testdata_heureka_clearcutting
}
\description{
Test data set for Brunberg's clearcutting model
}
\keyword{datasets}
