% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata_sensblt_v_Y_L_D}
\alias{testdata_sensblt_v_Y_L_D}
\title{Test data set for varying volume and terrain}
\format{
data.frame() with the following content:
\describe{
 \item{v}{numeric; mean stem volume m3fub}
 \item{Stems_ha}{Number of stems per hectare}
 \item{Nharv}{numeric; Number of harvested stems per hectare}
 \item{Y}{numeric; Surface structure (sv: Ytstruktur)}
 \item{L}{numeric; Slope class (sv: Lutning), Definition according “SkogForsk
 terrängtypschema” where 1 is flat (0-10\%), 2 is 10-20\%, 3 is 20-33\%, 4 is 33-50\%, 5 is >50\% }
 \item{D}{Terrain transport distance (one way average)}
}
}
\usage{
testdata_sensblt_v_Y_L_D
}
\description{
Test data set for varying volume and terrain
}
\keyword{datasets}
