% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heureka brunberg.R
\name{trunc_vls}
\alias{trunc_vls}
\title{trunc_vls}
\usage{
trunc_vls(x, min_val, max_val)
}
\arguments{
\item{max_val}{}
}
\value{
x truncated to within min_val and max_val
}
\description{
trunc_vls
}
\examples{
trunc_vls(1:10, 3, 6)
}
