% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunk_rng.R
\name{trunk_rng}
\alias{trunk_rng}
\title{trunk_rng}
\usage{
trunk_rng(x, min_val, max_val)
}
\arguments{
\item{x}{numeric vector to be trunkated}

\item{min_val}{minimum value, x lower than min_val will be replaced with min_val}

\item{max_val}{max value, x larger than min_val will be replaced with max_val}
}
\value{
x where extreme values ar truncated to within min_val and max_val
}
\description{
trunk_rng
}
\examples{
trunk_rng(1:10, 3, 6)
}
