% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heureka brunberg.R
\name{harv_forw}
\alias{harv_forw}
\title{Time consumption harvester and forwarder}
\usage{
harv_forw(stands_df, forw_size = "large", modelversion = "Talbot16")
}
\arguments{
\item{stands_df}{A data frame of stands for which harvester and forwarder
time consumption is to be estimated.


Mandatory variables in stands_df:
  v mean volume m3fub of the harvested trees in the stand
  Nharv Number of trees per hectare

 Optional variables, of which default values is set as first value:
  nbAssortments = 4; Number of qualities and species. Afect t7 ForwarderSortingTime
  treatment = one of "clearcutting"|"thinning"
  L = 2, slope class (1:5) where 5 is steep
  Y = 2, surface structure (1:5) where 5 is uneven
  D = 300, Terrain transport distance (one way average)}

\item{forw_size}{can be "small", "medium", "large"

Should also take "xl" and "xxl" but not yet implemented}

\item{modelversion}{one of "Brunberg" | "Talbot16"}
}
\value{
a data frame with harvester and forwarder time consumption (hrs hectare^-1)
\describe{
\item{harv_G15h.ha}{Harvester G15 hours per hectare}
\item{forw_G15h.ha}{Forwarder G15 hours per hectare}}
#' @export
}
\description{
Time consumption harvester and forwarder
}
\details{
\strong{variables in stands_df:}
\describe{
\item{v}{mean m3fub of harvested trees}
\item{Nharv}{N harvested trees per hectare}}
}
\examples{
harv_forw(stands_df = testdf_volume) \%>\% dplyr::glimpse()
harv_forw(stands_df = stand_data_norsk_prodnorm, forw_size = "large", modelversion = "Talbot16") \%>\% dplyr::glimpse()

}
